refactoring transactions table to own component
